Multi sound working now
