+++ TeslaGamesEngine/TeslaGamesEngine.cpp
@@ -136,13 +136,18 @@ int main()
 	ImGui_ImplOpenGL3_Init(glsl_version);
 
 	// Imgui setting END
-	AudioBoomBox* audioObject = new AudioBoomBox("finalCount.wav");
-	AudioBoomBox* audioObject2 = new AudioBoomBox("bounce.wav");
-	audioObject->initialize();
-	audioObject2->initialize();
-	audioObject->playSound();
-	audioObject2->playSound();
 
+	//Audio system setup
+	AudioEngine audioSystem = AudioEngine();
+	AudioBoomBox audioObject = audioSystem.createBoomBox(2);
+	AudioBoomBox audioObject2 = audioSystem.createBoomBox(3);
+
+	//The key is now that multiple sounds can be played at once. As long as sound card can support it
+	//Comment out one sound if you dont wanna hear it
+	audioObject.playSound();
+	audioObject2.playSound();
+
+	//End of audio system setup/demo
 	while (!mainWindow.getShouldClose())
 	{
 		
@@ -158,18 +163,10 @@ int main()
 		// Get + Handle User Input
 		glfwPollEvents();
 
-
-		
-
 		// Start the Dear ImGui frame
-		
 			ImGui_ImplOpenGL3_NewFrame();
 			ImGui_ImplGlfw_NewFrame();
 			ImGui::NewFrame();
-
-			
-
-		
 		{
 			static float f = 0.0f;
 			static int counter = 0;
@@ -210,13 +207,9 @@ int main()
 		meshList[1]->RenderMesh();
 
 
-
 		ImGui_ImplOpenGL3_RenderDrawData(ImGui::GetDrawData());
-
-			//
 		// imgui ends here
 
-		
 		// TODO: Load shader in a material struct in the model (Basically all of the following code refactored to being in model
 		mainWindow.swapBuffers();
 	}
